// src/app/shared/enum/shared-enum.ts
export enum MENU_OPTION {
  DELETE = 'DESTROY',
  EDIT = 'EDIT',
  MOVE_TO_TRASH = 'MOVE TO TRASH',
  RESTORE = 'RESTORE',
}

export enum MODE {
  VIEW = 'VIEW_ONLY',
  TRASH = 'TRASH',
  EDIT = 'EDIT',
}

export enum MODULE {
  COMMENT = 'COMMENT',
  POST = 'POST',
  MAIN_POST = 'MAIN_POST',
}

export enum ACTION {
  CREATE = 'Create',
  UPDATE = 'Update',
}